split api functional & start dev apiAuth
